Add restore_single action to backup and restore script
