+++ functions/backup_restore/main.py
@@ -6,7 +6,7 @@ from restore_manager import RestoreManager
 def main():
     parser = argparse.ArgumentParser(description="Backup and Restore Manager")
     parser.add_argument("dataset_abs_path", type=Path, help="The absolute path to the dataset parent (e.g., /mnt/POOL/DATASET).")
-    parser.add_argument("action", choices=["backup_all", "export", "restore_all", "import", "delete", "list"], help="Action to perform.")
+    parser.add_argument("action", choices=["backup_all", "export", "restore_all", "restore_single", "import", "delete", "list"], help="Action to perform.")
     parser.add_argument("backup_name", nargs="?", default=None, help="The name of the backup to perform the action on.")
     parser.add_argument("app_name", nargs="?", default=None, help="The name of the chart to import (only for 'import' action).")
     parser.add_argument("--retention", type=int, default=None, help="The number of backups to retain.")
@@ -25,11 +25,13 @@ def main():
             utility.backup_all(retention)
         elif action == "export":
             utility.export_chart_info(retention)
-    elif action in ["restore_all", "import"]:
+    elif action in ["restore_all", "restore_single", "import"]:
         utility = RestoreManager(dataset_abs_path)
         if backup_name:
             if action == "restore_all":
                 utility.restore_all(backup_name)
+            elif action == "restore_single":
+                utility.restore_single(backup_name, app_name)
             elif action == "import":
                 utility.import_chart(backup_name, app_name)
         else:
